+++ source/gui/addonStoreGui/viewModels/store.py
@@ -262,6 +262,13 @@ class AddonStoreVM:
 				actionTarget=selectedListItem,
 			),
 			AddonActionVM(
+				# Translators: Label for an action that opens the VirusTotal scan results for the selected addon
+				displayName=pgettext("addonStore", "VirusTotal scan results"),
+				actionHandler=lambda aVM: startfile(cast(_AddonStoreModel, aVM.model).scanResults.scanUrl),
+				validCheck=lambda aVM: isinstance(aVM.model, _AddonStoreModel)
+				and aVM.model.scanResults is not None,
+			),
+			AddonActionVM(
 				# Translators: Label for an action that shows changelog for the selected addon
 				displayName=pgettext("addonStore", "&What's new"),
 				actionHandler=lambda aVM: ui.browseableMessage(